webnn/div.html

<!doctype html>
<meta charset=utf-8>
<title>test element-wise binary operations</title>
<link rel="help" href="https://webmachinelearning.github.io/webnn/#api-mlgraphbuilder-binary">
<script src=../resources/testharness.js></script>
<script src=../resources/testharnessreport.js></script>
<script src="./webnn-polyfill.js"></script>
<div id=log></div>
<script type="module">
  'use strict';
  import * as utils from './utils.js';
  let builder;
  setup(() => {
    const context = navigator.ml.createContext();
    builder = new MLGraphBuilder(context);
  })

  test(() => {
    const a = builder.input('a', {type: 'float32', dimensions: [2,3,4]});
    const b = builder.constant(
        {type: 'float32', dimensions: [2,3,4]}, new Float32Array([
        0.45513507393113173,
        0.9329701166765885,
        0.45717846572633336,
        0.7784609033697032,
        0.6992337709529468,
        0.47849715417983885,
        0.6939645376296133,
        0.01634559778584821,
        0.8819085752111917,
        0.6975036074364973,
        0.861587233775087,
        0.2785635186894868,
        0.6801561531241653,
        0.21285450136842465,
        0.37051782325052995,
        0.6237038332271383,
        0.440127455565954,
        0.0656937360000065,
        0.5208334797791523,
        0.5739108358675569,
        0.11310017634727498,
        0.29064712136538584,
        0.33900594727817257,
        0.019759666931747466, 
        ]));
    const c = builder.div(a, b);
    const graph = builder.build({c});
    const inputs = {
      'a': new Float32Array([
      0.04447953887691036,
        0.9902262167146556,
        0.18841561960300113,
        0.9195430456388485,
        0.34455126683999504,
        0.8578780185454358,
        0.7831234092952899,
        0.048357024567023066,
        0.8436948153683268,
        0.8821329149911432,
        0.9392319649315293,
        0.5940836117208761,
        0.9842279644548573,
        0.9559909810057179,
        0.030384067698445705,
        0.11376630439761004,
        0.3807257989823989,
        0.3846086403891549,
        0.16844453295498596,
        0.664698945014671,
        0.6813208085328701,
        0.006303789966506157,
        0.7639197615904112,
        0.06234929616653395, 
      ]),
    };
    const outputs = {c: new Float32Array(utils.sizeOfShape([2,3,4]))};
    graph.compute(inputs, outputs);
    const expected = [
        0.09772821613752566, 
        1.0613697041466064, 
        0.41212706574807606, 
        1.1812321487931465, 
        0.4927554719939017, 
        1.792859186416415, 
        1.1284775616492135, 
        2.958412729872131, 
        0.9566692501728835, 
        1.2647001471909305, 
        1.090118247012828, 
        2.1326683928885095, 
        1.4470617665281673, 
        4.49128853211808, 
        0.08200433499227691, 
        0.1824043694087399, 
        0.8650353304881393, 
        5.854570980543972, 
        0.3234134123374923, 
        1.1581920108022938, 
        6.024047269748451, 
        0.021688809222993724, 
        2.253411091232491, 
        3.155381939477865,
    ];
    assert_array_approx_equals_ulp('div', outputs.c, expected, 0);
  }, 'div 2**-126  / 1');
  test(() => {
    const a = builder.input('a', {type: 'float32', dimensions: [2,3,4]});
    const b = builder.constant(
        {type: 'float32', dimensions: [2,3,4]}, new Float32Array([
        1.7724128920039317e-38,
        2.0147313213775673e-38,
        2.1341711020272885e-38,
        2.166418313143517e-38,
        1.977705724799706e-38,
        1.658251770084896e-38,
        1.9128086136940154e-38,
        1.371341209669247e-38,
        1.7013959151651063e-38,
        2.275140299640556e-38,
        1.5716740483302364e-38,
        1.8481539310087546e-38,
        1.9149823661423905e-38,
        1.899663035400152e-38,
        1.2478725657276344e-38,
        1.2021046713294018e-38,
        1.6746367799213846e-38,
        2.2557044010027228e-38,
        2.0106983910850034e-38,
        2.0550748065889805e-38,
        2.2201624840128717e-38,
        2.2086671877680075e-38,
        1.2074153717510325e-38,
        1.3471743985533825e-38,
        ]));
    const c = builder.div(a, b);
    const graph = builder.build({c});
    const inputs = {
      'a': new Float32Array([
      2.1996348240699856e-38,
        2.1681148969856212e-38,
        2.0627696018202352e-38,
        1.662575810762009e-38,
        1.4099993499475046e-38,
        1.3310117817263286e-38,
        1.9369397177469903e-38,
        1.894728457510042e-38,
        1.9433629760569645e-38,
        1.5064113266754953e-38,
        2.2955714268944255e-38,
        1.9081046985712263e-38,
        1.696744885793692e-38,
        2.0570529411472357e-38,
        1.4205584682834975e-38,
        1.27369084854511e-38,
        1.812036233448369e-38,
        1.6736363300034075e-38,
        1.6921579413306318e-38,
        1.4620578037497756e-38,
        1.6685944853973987e-38,
        1.6839736080410745e-38,
        1.898158345727739e-38,
        1.2535916410215008e-38,
      ]),
    };
    const outputs = {c: new Float32Array(utils.sizeOfShape([2,3,4]))};
    graph.compute(inputs, outputs);
    const expected = [
    1.241039734022148, 
        1.0761310324511053, 
        0.9665436852091064, 
        0.767430648400298, 
        0.7129469932086604, 
        0.8026596477917138, 
        1.0126155350201884, 
        1.381660847169488, 
        1.142216787248121, 
        0.6621179919820723, 
        1.4605900182250038, 
        1.0324381895666825, 
        0.8860368198646529, 
        1.0828514861920924, 
        1.1383842447527244, 
        1.0595507021335682, 
        1.082047316274419, 
        0.7419572924800918, 
        0.841577209607015, 
        0.7114377535368183, 
        0.751564129838582, 
        0.7624388216419479, 
        1.5720839655825891, 
        0.9305340439720556, 
    ];
    assert_array_approx_equals_ulp('div', outputs.c, expected, 0);
  }, 'div 2**-126  / 2');
  test(() => {
    const a = builder.input('a', {type: 'float32', dimensions: [2,3,4]});
    const b = builder.constant(
        {type: 'float32', dimensions: [2,3,4]}, new Float32Array([
        4.760897227875005e+37,
        5.086196241960284e+37,
        7.681770910676816e+37,
        5.256550391433613e+37,
        8.26351865343929e+37,
        8.125206882375855e+37,
        6.939931050425726e+37,
        4.265590664146158e+37,
        5.0381432212745435e+37,
        4.392348561511389e+37,
        8.256548516600667e+37,
        5.285473627655199e+37,
        5.958109795516231e+37,
        6.896904520158982e+37,
        5.01850238816788e+37,
        7.595413270213756e+37,
        6.650334617764448e+37,
        6.0812050013736265e+37,
        4.313447370221653e+37,
        7.235529492851894e+37,
        5.227378602625893e+37,
        7.297682088260833e+37,
        5.947063584829604e+37,
        7.994749441150284e+37,
        ]));
    const c = builder.div(a, b);
    const graph = builder.build({c});
    const inputs = {
      'a': new Float32Array([
        1.990125004777618e-38,
        1.2368178039825982e-38,
        2.1684250701436127e-38,
        1.7995849477418025e-38,
        1.7902477499423783e-38,
        1.541270146658142e-38,
        1.7623223738028622e-38,
        1.8658308179131031e-38,
        1.3563885032755036e-38,
        1.9760665502986022e-38,
        1.1774482564353866e-38,
        1.955458481695229e-38,
        2.259555836661282e-38,
        2.0557997120525696e-38,
        2.267219604047388e-38,
        2.142590607323001e-38,
        1.4523246476426738e-38,
        2.182090927886304e-38,
        1.4800928667365876e-38,
        2.3455245351638445e-38,
        1.7728212571642427e-38,
        2.1003008498863792e-38,
        1.6262672864499002e-38,
        1.4755640479519735e-38,
      ]),
    };
    const outputs = {c: new Float32Array(utils.sizeOfShape([2,3,4]))};
    graph.compute(inputs, outputs);
    const expected = [
        4.1801469544972663e-76, 
        2.4317146746699518e-76, 
        2.822819237071676e-76, 
        3.423509362098979e-76, 
        2.1664472787234216e-76, 
        1.8968995731065818e-76, 
        2.5393946438340384e-76, 
        4.374144086529657e-76, 
        2.69223887393254e-76, 
        4.4988837352622257e-76, 
        1.4260780446793257e-76, 
        3.699684492726778e-76, 
        3.792403823040838e-76, 
        2.980757100585729e-76, 
        4.517721480800349e-76, 
        2.820900629232914e-76, 
        2.1838369512463446e-76, 
        3.588254182178387e-76, 
        3.4313456029499e-76, 
        3.2416764211672816e-76, 
        3.391415453002951e-76, 
        2.878038292823082e-76, 
        2.734571882833663e-76, 
        1.845666407451125e-76,
    ];
    assert_array_approx_equals_ulp('div', outputs.c, expected, 0);
  }, 'div 2**-126  / 3');

  test(() => {
    const a = builder.input('a', {type: 'float32', dimensions: [2,3,4]});
    const b = builder.constant(
        {type: 'float32', dimensions: [2,3,4]}, new Float32Array([
        5.892008280194684e+37,
        5.555377151629954e+37,
        6.498379415955097e+37,
        7.777097337688773e+37,
        4.399735517921792e+37,
        5.420402270840776e+37,
        7.171471060050253e+37,
        5.815263820835398e+37,
        6.473425864648833e+37,
        5.172310538755502e+37,
        4.618496159499228e+37,
        5.394288563835964e+37,
        6.774711074404044e+37,
        6.591902469816717e+37,
        6.208568000392576e+37,
        5.008269399395899e+37,
        5.627531022226794e+37,
        5.155579758190636e+37,
        6.518924611127592e+37,
        5.829823129470408e+37,
        4.596835613632579e+37,
        5.733775371988708e+37,
        6.685273810546262e+37,
        5.574536877663488e+37,
        ]));
    const c = builder.div(a, b);
    const graph = builder.build({c});
    const inputs = {
      'a': new Float32Array([
        5.859722373622668e+37,
        4.731647438004698e+37,
        4.37100453340166e+37,
        4.823497862326285e+37,
        5.114946891747196e+37,
        4.692385177448277e+37,
        7.706391735109153e+37,
        7.86684189691531e+37,
        8.227996021809832e+37,
        8.285156818395987e+37,
        5.36469905038862e+37,
        5.820965827563686e+37,
        5.207152597607143e+37,
        4.887260485740213e+37,
        8.309288398426632e+37,
        5.517202105429841e+37,
        8.03469014761029e+37,
        4.331909776254237e+37,
        4.508750198801839e+37,
        5.761803249161694e+37,
        7.38352164120521e+37,
        4.257163129808179e+37,
        6.193361442541632e+37,
        6.507324606896089e+37,
      ]),
    };
    const outputs = {c: new Float32Array(utils.sizeOfShape([2,3,4]))};
    graph.compute(inputs, outputs);
    const expected = [
        0.9945203901561812, 
        0.8517238900002364, 
        0.6726299364222692, 
        0.6202182707616399, 
        1.1625578107847796, 
        0.8656894715528975, 
        1.0745900904542103, 
        1.3527919178368748, 
        1.271041978984057, 
        1.6018289614121777, 
        1.1615683688194949, 
        1.0790979679115098, 
        0.7686161875272599, 
        0.7414036400141247, 
        1.3383582813140205, 
        1.1016184764532295, 
        1.4277469312698683, 
        0.8402371759203532, 
        0.6916401811282723, 
        0.9883324281375081, 
        1.60621833404447, 
        0.742471208517473, 
        0.9264185159882876, 
        1.1673300849385662,
    ];
    assert_array_approx_equals_ulp('div', outputs.c, expected, 0);
  }, 'div 2**126  / 1');

  test(() => {
    const a = builder.input('a', {type: 'float32', dimensions: [2,3,4]});
    const b = builder.constant(
        {type: 'float32', dimensions: [2,3,4]}, new Float32Array([
        1.6087448841071964e-38,
        2.1633317321541952e-38,
        2.0073498536858038e-38,
        1.257265364393177e-38,
        1.563754490008614e-38,
        1.17862082610584e-38,
        1.7297456127945125e-38,
        1.781993840256496e-38,
        1.8744082700217967e-38,
        1.6856882672329858e-38,
        1.7279356096595697e-38,
        1.6374634216602454e-38,
        1.997996670120333e-38,
        1.7004160819633032e-38,
        1.9112306083608425e-38,
        1.3060582059289934e-38,
        1.2309463498825027e-38,
        1.2558648444217821e-38,
        1.8562527872449883e-38,
        1.433645161154843e-38,
        2.2763365248771767e-38,
        1.721513328540402e-38,
        2.0291711385347364e-38,
        1.1879531278103388e-38,
        ]));
    const c = builder.div(a, b);
    const graph = builder.build({c});
    const inputs = {
      'a': new Float32Array([
        6.394989610922957e+37,
        4.297340860822846e+37,
        7.953515890672566e+37,
        7.038124663535706e+37,
        4.315040292843341e+37,
        6.575527878871825e+37,
        4.503566836945878e+37,
        6.560167436128598e+37,
        8.413405399235992e+37,
        8.087152744507112e+37,
        6.990015922358865e+37,
        6.833955015148073e+37,
        5.035281425064017e+37,
        4.925026141214545e+37,
        6.712230504973476e+37,
        7.116345099788938e+37,
        6.628708106053482e+37,
        6.063373771881942e+37,
        5.3188117516987225e+37,
        5.078750410053795e+37,
        5.064682602146651e+37,
        7.460133799806623e+37,
        5.174219034491079e+37,
        7.312918774864596e+37,
      ]),
    };
    const outputs = {c: new Float32Array(utils.sizeOfShape([2,3,4]))};
    graph.compute(inputs, outputs);
    const expected = [
        3.975142158398831e+75, 
        1.9864456278019155e+75, 
        3.9621971606338e+75, 
        5.5979627394990534e+75, 
        2.7594103296992427e+75, 
        5.5790019429720675e+75, 
        2.603600670314801e+75, 
        3.681363699430265e+75, 
        4.488566089786915e+75, 
        4.797537541019946e+75, 
        4.045298842898438e+75, 
        4.173500870156255e+75, 
        2.520165073528755e+75, 
        2.8963653034427327e+75, 
        3.5119940396570917e+75, 
        5.4487197182204556e+75, 
        5.3850503774483846e+75, 
        4.828046424592452e+75, 
        2.8653488297748455e+75, 
        3.5425435440124616e+75, 
        2.2249270030141628e+75, 
        4.33347431944181e+75, 
        2.549917518651177e+75, 
        6.1558984135543524e+75,
    ];
    assert_array_approx_equals_ulp('div', outputs.c, expected, 0);
  }, 'div 2**126  / 2');
</script>
